+++ src/main/java/chess/Application.java
@@ -1,15 +1,6 @@
 package chess;
 
 import chess.controller.ChessGameController;
-import chess.domain.Score;
-import chess.domain.Team;
-import chess.domain.state.BoardInitialize;
-import chess.domain.state.GameState;
-import chess.domain.state.WhiteTurn;
-import chess.view.InputView;
-import chess.view.OutputView;
-
-import java.util.List;
 
 public class Application {
     public static void main(String[] args) {
