+++ src/main/java/chess/domain/Score.java
@@ -1,6 +1,9 @@
 package chess.domain;
 
 import chess.domain.piece.Piece;
+import chess.domain.position.Column;
+import chess.domain.position.Position;
+import chess.domain.position.Row;
 
 import java.util.EnumMap;
 import java.util.Map;
