+++ src/main/java/chess/domain/piece/Direction.java
@@ -1,6 +1,5 @@
-package chess.domain;
+package chess.domain.piece;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.function.BiPredicate;
