+++ src/main/java/chess/domain/state/BlackTurn.java
@@ -1,8 +1,8 @@
 package chess.domain.state;
 
-import chess.domain.Position;
 import chess.domain.Team;
 import chess.domain.piece.Piece;
+import chess.domain.position.Position;
 
 import java.util.Map;
