+++ src/main/java/chess/domain/state/BoardInitialize.java
@@ -1,10 +1,10 @@
 package chess.domain.state;
 
-import chess.domain.Column;
-import chess.domain.Position;
-import chess.domain.Row;
 import chess.domain.Team;
 import chess.domain.piece.*;
+import chess.domain.position.Column;
+import chess.domain.position.Position;
+import chess.domain.position.Row;
 
 import java.util.HashMap;
 import java.util.Map;
