+++ src/main/java/chess/domain/state/GameState.java
@@ -1,7 +1,8 @@
 package chess.domain.state;
 
-import chess.domain.*;
+import chess.domain.Team;
 import chess.domain.piece.Piece;
+import chess.domain.position.Position;
 
 import java.util.Map;
