+++ src/test/java/chess/domain/KnightTest.java
@@ -1,6 +1,7 @@
 package chess.domain;
 
 import chess.domain.piece.Knight;
+import chess.domain.position.Position;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
