+++ src/test/java/chess/domain/state/WhiteTurnTest.java
@@ -1,6 +1,6 @@
 package chess.domain.state;
 
-import chess.domain.Position;
+import chess.domain.position.Position;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
